wado-compiler: extract lift submodule from cm_binding
Move the memory-based CM lift (synthesize_lift, the WASI-aware
synthesize_lift_with_context, and their list/option/result/tuple/
variant/enum/struct workers) into `synthesis/cm_binding/lift.rs`.
The `synthesize_lift_flat_result` helper that lives with import-adapter
synthesis still needs `try_lift_wasi_variant_or_enum`, so it's exposed
as `pub(super)`.

Also drops two stale doc comments left behind by the earlier types
extraction, where the `kebab_to_pascal` and `is_unit_type` doc lines
had drifted onto the wrong functions.
